feature: change password

+++ Infrastructure/Repositories/SQLServer/DataContext/ExamDetail.cs
@@ -10,7 +10,7 @@ namespace Infrastructure.Repositories.SQLServer.DataContext
     {
         public required Guid ExamId { get; set; }
         public required Guid QuestionId { get; set; }
-        public required int Score { get; set; }
+        public required double Score { get; set; }
         public Exam Exam { get; set; } = null!;
         public Question Question { get; set; } = null!;
     }
