+++ Infrastructure/Repositories/SQLServer_Read/DataContext/ExamDetail.cs
@@ -10,6 +10,6 @@ namespace Infrastructure.Repositories.SQLServer_Read.DataContext
     {
         public required Guid ExamId { get; set; }
         public required Guid QuestionId { get; set; }
-        public required int Score { get; set; }
+        public required double Score { get; set; }
     }
 }
